+++ lib/src/runtime/root/root_indicator.dart
@@ -1,8 +1,8 @@
 enum RootIndicatorType {
   suBinary,
   suExecution,
-  testKeys,
   dangerousProps,
   writableSystem,
   knownRootApp,
+  testKeys,
 }